Link filtering for admins

+++ src/filters/LinkFilter.js
@@ -11,6 +11,12 @@ class LinkFilter {
         if(this.m.author.bot)
             return;
 
+        let mutedRole = this.m.member.roles.find(role => role.name === "admin");
+
+        if(mutedRole)
+            return;
+
+
         if (this.message.includes('https://') || this.message.includes('http://') || this.message.includes('www.')) {
             this.m.delete();
             console.log('Link found deleting');
